backend/requirements.txt
# FastAPI 서버
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.1
httpx>=0.27.0

# LangChain
langchain-classic>=1.0.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0

# OpenAI
openai>=1.40.0

# Supabase pgvector
supabase>=2.5.0

# BM25 (Hybrid Search)
rank-bm25>=0.2.2

# Redis 캐시
redis>=5.0.0

# 유틸리티
loguru>=0.7.2
